+++ tests/FunctionalTest/Fixtures/config/di2.php
@@ -8,6 +8,5 @@ use function DI\get;
 use function DI\create;
 
 return [
-    ContainerAwareController::class => create(),
-    Class1::class => create()->constructor(get('class2')),
+    Class3::class => create()->constructor(get('class2_import')),
 ];